Bugfix: package installation using devtools
Removed compiled source files and added a .gitignore file so that
they are not re-added to the repository in the future. When present,
these files prevent R CMD build / install.packages from being able
to install the package on machines with different architectures to
the package author/compiler's machine. Removing these files means
that the package can be installed by anyone using
devtools::install_bitbucket()